feat: 컴포넌트화 (#97)
